fix the errors with pylint
